Warnings… helper functions.

+++ libraries/jlewallen/WireHelpers.cpp
@@ -14,7 +14,7 @@ uint8_t wireWaitForAvailableBytes(uint8_t bytes, uint16_t time) {
 uint8_t wireReadBlock(void *p, size_t sz, uint16_t time) {
   uint8_t *ptr = (uint8_t *)p;
   uint32_t started = millis();
-  while (ptr - (uint8_t *)p < sz) {
+  while ((size_t)(ptr - (uint8_t *)p) < sz) {
     if (Wire.available()) {
       *ptr++ = Wire.read();
     }
